refactor(compiler): decompose cfl.py (1146 to 656 lines)
Phase 5.2: extract two cohesive method groups from the CFLPlanner
gravity-well into sibling modules, keeping CFLPlanner as a thin facade
(one-line delegators). Pure code movement; generated SQL byte-identical
(no drift-snapshot changes).

- compiler/cfl_exclude.py: dimensionsExclude anti-join group
  (plan_dimensions_exclude, partition_dimensions, build_group_distinct_select,
  build_existing_pairs_select)
- compiler/cfl_projection.py: outer-projection / null-padding / measure-ref
  helpers (group_measures_by_object, group_dimensions_into_legs,
  build_outer_metric_expr, substitute_outer_refs, remap_cfl_order_by, etc.)

plan(), _validate_fanout(), and the core _plan_union_all() stay in cfl.py.
Stateful helpers take the planner as first arg; CFLPlanner reference is
TYPE_CHECKING-only to keep 0 import cycles.

Full suite green (2485 passed); ruff/format/mypy clean; 0 import cycles.
